Styling tweaks and bug fixes
I spotted th FAQ page contact form heading was hidden as it inherited the same colour as $env. So I dumplicated the include and called it -faq. This meant altering the faq.html and sass.

Also tweaked the footer_redesign.html to add back in the phone number as we now have one.

Also made all the text opd-blue.